# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "RegisterRequest": {
        "additionalProperties": false,
        "description": "§1's registration body.\n\n`extra=\"forbid\"` so a client sending `role: \"operator\"` gets a 422 rather\nthan having it ignored - an ignored field is a field someone will one day\nwire up by accident.\n\nThe password bounds are declared here *and* enforced by\n`passwords.check_length` in the service. Not redundant: pydantic's error is\na 422 `validation_error` naming the field, and §1 wants\n`password_too_short` / `password_too_long` specifically\n(`AC-AUTH-01.1`), which only the service can raise. The constraint here is\na cheap outer bound that stops a 10 MB body reaching Argon2; the service is\nwhat produces the specified error code.",
        "properties": {
          "consent_items": {
            "items": {
              "type": "string"
            },
            "title": "Consent Items",
            "type": "array"
          },
          "consent_version": {
            "title": "Consent Version",
            "type": "integer"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          },
          "tz": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tz"
          }
        },
        "required": [
          "email",
          "password",
          "consent_version"
        ],
        "title": "RegisterRequest",
        "type": "object"
      },
      "RegisterResponse": {
        "description": "§1's 202 body.\n\nOne field, and deliberately uninformative. `AC-AUTH-01.4` requires the\nresponse for an already-registered address to be *byte-identical* to a fresh\none, so there is nothing here that could differ between the two - no user\nid, no \"created\" flag, no email echo. Anything that varied would be the\nenumeration oracle §1 exists to close.",
        "properties": {
          "detail": {
            "default": "Check your email to finish signing up.",
            "title": "Detail",
            "type": "string"
          }
        },
        "title": "RegisterResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "JobPilot API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "description": "§1's registration. Always 202, whether or not the address was taken.\n\n`AC-AUTH-01.4` requires the two responses to be byte-identical, so the\nstatus and the body are fixed here and the service returns nothing to vary\nthem with.",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register an account",
        "tags": [
          "auth"
        ]
      }
    }
  }
}